update schemas to v3
