+++ studio/schemas/documents/book.ts
@@ -1,28 +1,30 @@
-export default {
+import { defineField, defineType } from "sanity";
+
+export default defineType({
 	name: "book",
 	title: "Knyga",
 	type: "document",
 	fields: [
-		{
+		defineField({
 			name: "title",
 			title: "Sekcijos pavadinimas",
 			type: "localeString",
-		},
-		{
+		}),
+		defineField({
 			type: "a11yImage",
 			name: "cover",
 			title: "Viršelio nuotrauka",
-		},
-		{
+		}),
+		defineField({
 			name: "description",
 			title: "Aprašymas",
 			type: "localeEditorial",
-		},
-		{
+		}),
+		defineField({
 			name: "cta",
 			title: "Nuoroda",
 			type: "localeCta",
-		}
+		}),
 	],
 	preview: {
 		select: {
@@ -30,5 +32,4 @@ export default {
 			media: "cover.1.image",
 		},
 	},
-};
-
+});
