+++ studio/schemas/documents/category.ts
@@ -1,0 +1,19 @@
+import { defineField, defineType } from "sanity";
+
+export default defineType({
+	name: "category",
+	title: "Category",
+	type: "document",
+	fields: [
+		defineField({
+			name: "title",
+			title: "Title",
+			type: "string",
+		}),
+		defineField({
+			name: "description",
+			title: "Description",
+			type: "text",
+		}),
+	],
+});
