--- studio/schemas/documents/config.js
@@ -1,43 +1,0 @@
-export default {
-  name: "config",
-  title: "Config",
-  type: "document",
-  fields: [
-    {
-      title: "SEO Metadata",
-      name: "seo",
-      type: "object",
-      fields: [
-        {
-          name: "title",
-          title: "Title",
-          type: "string",
-        },
-        {
-          name: "description",
-          title: "Description",
-          type: "text",
-        },
-        {
-          name: "image",
-          title: "Image",
-          type: "a11yImage",
-        },
-        {
-          name: "keywords",
-          title: "Keywords",
-          type: "array",
-          of: [{ type: "string" }],
-          options: {
-            layout: "tags",
-          },
-        },
-      ],
-    },
-  ],
-  preview: {
-    select: {
-      title: 'seo.title'
-    }
-  }
-};
