+++ studio/schemas/documents/consultant.ts
@@ -1,27 +1,28 @@
-export default {
+import { defineField, defineType } from "sanity";
+
+export default defineType({
 	name: "consultant",
 	title: "Konsultantė / Konsultantas",
 	type: "document",
 	fields: [
-		{
+		defineField({
 			name: "order",
 			title: "Order",
 			type: "number",
 			hidden: true,
-		},
-		{
+		}),
+		defineField({
 			name: "name",
 			title: "Vardas (+ titulas)",
 			type: "localeString",
-		},
-		{
+		}),
+		defineField({
 			name: "headline",
 			title: "Pozicija / Apibūdinimas",
 			type: "localeString",
-			description:
-				"Pvz.: pscichoterapeutė (-as), mokymų vadovė (-as) ir pan.",
-		},
-		{
+			description: "Pvz.: pscichoterapeutė (-as), mokymų vadovė (-as) ir pan.",
+		}),
+		defineField({
 			name: "slug",
 			title: "Slug",
 			type: "slug",
@@ -31,23 +32,23 @@ export default {
 				source: "name",
 				maxLength: 96,
 			},
-		},
-		{
+		}),
+		defineField({
 			type: "a11yImage",
 			name: "profile",
 			title: "Profilio nuotrauka",
-		},
-		{
+		}),
+		defineField({
 			name: "shortBio",
 			title: "Trumpas aprašymas",
 			type: "localeText",
 			description: "Matysis ant pirmojo puslapio.", // TODO: add character limit
-		},
-		{
+		}),
+		defineField({
 			name: "bio_new",
 			title: "Pilnas aprašymas",
 			type: "localizedConsultantBioSections",
-		},
+		}),
 	],
 	preview: {
 		select: {
@@ -55,4 +56,4 @@ export default {
 			media: "profile.1.image",
 		},
 	},
-};
+});
