+++ studio/schemas/documents/paslaugos.ts
@@ -1,6 +1,7 @@
 import { baseLanguage } from "../languages";
+import { defineField, defineType } from "sanity";
 
-export default {
+export default defineType({
 	type: "document",
 	name: "paslaugos",
 	title: "Paslaugos",
@@ -10,55 +11,60 @@ export default {
 		},
 	},
 	fields: [
-		{
+		defineField({
 			name: "order",
 			title: "Order",
 			type: "number",
 			hidden: true,
-		},
-		{
+		}),
+		defineField({
 			name: "title",
 			title: "Title",
 			type: "localeString",
-		},
-		{
+		}),
+		defineField({
 			name: "slug",
 			title: "Slug",
 			type: "localeSlug",
 			description:
 				"Pagal jį bus sukurta nuoroda (URL). Pvz: /konsultacijos-porai",
-		},
-		{
+		}),
+		defineField({
 			name: "truncatedBlurb",
 			title: "Prierašas",
 			type: "localeText", // TODO: validation for character limit
-		},
-		{
+		}),
+		defineField({
 			name: "blurb",
 			title: "Pilnas aprašas",
 			type: "localeEditorial",
-		},
-		{
+		}),
+		defineField({
 			name: "priceBlurb",
 			title: "Kainodaros prierašas",
 			type: "localeEditorial",
-		},
-		{
+		}),
+		defineField({
 			name: "kainos",
 			title: "Kainos",
 			description: "Kainos (tik porų konsultacijoms kol kas)",
 			type: "array",
 			of: [{ type: "price" }],
+			//HACK: show prices only on certain pages
 			hidden: ({ document }) => {
-				if (
-					document.slug.lt.current == "poros-psichoterapija" ||
-					document.slug.en.current == "couple-psychotherapy"
-				) {
-					return false;
-				} else {
-					return true;
-				}
+				if (document?.slug) {
+					if (
+						//@ts-ignore
+						document?.slug.lt.current === "poros-psichoterapija" ||
+						//@ts-ignore
+						document?.slug.en.current === "couple-psychotherapy"
+					) {
+						return false;
+					} else {
+						return true;
+					}
+				} else { return false }
 			},
-		},
+		}),
 	],
-};
+});
