--- studio/schemas/documents/post.js
@@ -1,65 +1,0 @@
-export default {
-  name: "post",
-  title: "Post",
-  type: "document",
-  fields: [
-    {
-      name: "title",
-      title: "Title",
-      type: "string",
-    },
-    {
-      name: "slug",
-      title: "Slug",
-      type: "slug",
-      options: {
-        source: "title",
-        maxLength: 96,
-      },
-    },
-    {
-      name: "author",
-      title: "Author",
-      type: "reference",
-      to: { type: "consultant" },
-    },
-    {
-      name: "mainImage",
-      title: "Main image",
-      type: "image",
-      options: {
-        hotspot: true,
-      },
-    },
-    {
-      name: "categories",
-      title: "Categories",
-      type: "array",
-      of: [{ type: "reference", to: { type: "category" } }],
-    },
-    {
-      name: "publishedAt",
-      title: "Published at",
-      type: "datetime",
-    },
-    {
-      name: "body",
-      title: "Body",
-      type: "blockEditorial",
-    },
-  ],
-
-  preview: {
-    select: {
-      title: "title",
-      author: "consultant.name",
-      media: "mainImage",
-    },
-    prepare(selection) {
-      const { consultant } = selection;
-      return Object.assign({}, selection, {
-        subtitle: consultant && `by ${consultant}`,
-      });
-    },
-  },
-};
