+++ studio/schemas/documents/pricing.ts
@@ -1,6 +1,7 @@
 import { baseLanguage } from "../languages";
+import { defineField, defineType } from "sanity";
 
-export default {
+export default defineType({
 	type: "document",
 	name: "price",
 	title: "Kaina",
@@ -10,21 +11,21 @@ export default {
 		},
 	},
 	fields: [
-		{
+		defineField({
 			name: "label",
 			title: "Pajamų lygis",
 			type: "localeString",
 			description: "pvz.: iki 1000 €",
-		},
-		{
+		}),
+		defineField({
 			name: "price_1",
 			title: "Kaina kai 1 psichoterapeutas (EUR per val.)",
 			type: "number",
-		},
-		{
+		}),
+		defineField({
 			name: "price_2",
 			title: "Kaina kai 2 psichoterapeutai (EUR per val.)",
 			type: "number",
-		},
+		}),
 	],
-};
+});
