--- studio/schemas/languages.js
@@ -1,6 +1,0 @@
-export const languages = [
-  { id: "lt", title: "Lietuvių", isDefault: true },
-  { id: "en", title: "English" },
-];
-
-export const baseLanguage = languages.find((l) => l.isDefault);
