--- studio/schemas/objects/blockEditorial.js
@@ -1,16 +1,0 @@
-export default {
-  name: "blockEditorial",
-  title: "Editorial Content",
-  type: "object",
-  fields: [
-    {
-      type: "array",
-      name: "content",
-      of: [
-        {
-          type: "block",
-        },
-      ],
-    },
-  ],
-};
