+++ studio/schemas/objects/consultantBioSection.ts
@@ -1,19 +1,21 @@
-export default {
+import { defineArrayMember, defineField, defineType } from "sanity";
+
+export default defineType({
 	name: "consultantBioSection",
 	title: "Biography Section",
 	type: "object",
 	fields: [
-		{
+		defineField({
 			name: "title",
 			title: "Section title",
 			type: "string",
-		},
-		{
+		}),
+		defineField({
 			name: "body",
 			title: "Section body",
 			type: "array",
 			of: [
-				{
+				defineArrayMember({
 					type: "block",
 					styles: [{ title: "Normal", value: "normal" }],
 					lists: [{ title: "Bullet", value: "bullet" }],
@@ -23,8 +25,8 @@ export default {
 							{ title: "Emphasis", value: "em" },
 						],
 					},
-				},
+				}),
 			],
-		},
+		}),
 	],
-};
+});
