--- studio/schemas/objects/localizedCta.js
@@ -1,27 +1,0 @@
-import { languages } from "../languages";
-
-function createCta() {
-	let fieldsobj = [{
-		type: "string",
-		name: "url",
-		title: "URL nuoroda",
-		description: "Kur nukreipti vartotoją (pilnas su https://)"
-	}]
-
-	languages.map((lang) => {
-		fieldsobj.push({
-			type: "string",
-			name: lang.id,
-			title: lang.title,
-			description: "Nuorodos tekstas"
-		})
-	})
-	return fieldsobj
-}
-
-export default {
-	title: "Localized CTA",
-	name: "localeCta",
-	type: "object",
-	fields: createCta()
-};
