+++ studio/schemas/schema.ts
@@ -1,0 +1,40 @@
+import consultant from "./documents/consultant";
+import category from "./documents/category";
+import config from "./documents/config";
+import book from "./documents/book";
+import landingPage from "./documents/landingPage";
+import model from "./documents/model";
+import paslaugos from "./documents/paslaugos";
+import pricing from "./documents/pricing";
+
+// Objects
+import localizedString from "./objects/localizedString";
+import localizedSlug from "./objects/localizedSlug";
+import localizedCta from "./objects/localizedCta";
+import localizedText from "./objects/localizedText";
+import localizedEditorial from "./objects/localizedEditorial";
+import localizedConsultantBioSections from "./objects/localizedConsultantBioSections";
+import consultantBioSection from "./objects/consultantBioSection";
+import blockEditorial from "./objects/blockEditorial";
+import a11yImage from "./objects/a11yImage";
+
+export default [
+	book,
+	consultant,
+	category,
+	config,
+	model,
+	paslaugos,
+	pricing,
+	landingPage,
+	// Objects
+	a11yImage,
+	blockEditorial,
+	consultantBioSection,
+	localizedEditorial,
+	localizedConsultantBioSections,
+	localizedText,
+	localizedString,
+	localizedCta,
+	localizedSlug,
+];
